pyproject.toml
[project]
name = "pizza-shop-env"
version = "1.0.0"
description = "OpenEnv pizza shop operations environment with graded task ladder"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "openai>=1.50.0",
    "gradio>=4.44.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
